Moved the files into a folder in components called crud and moved login and signup into pages
