+++ src/components/crud/view.jsx
@@ -1,7 +1,7 @@
 import { Card, Title, Text, Badge, Button, Group, Space } from '@mantine/core';
 import React, { useEffect, useState } from "react";
 import Read from "../components/read"
-import Update from '../components/update';
+import Update from './update';
 
 
 export default function View(props) {
